project/requirements.txt
dash==2.18.2
faiss_cpu==1.9.0.post1
holidays==0.63
lightgbm==4.5.0
mlflow_skinny==2.19.0
mlforecast==1.0.0
numpy==1.26.4
openai==0.28.0
optuna==4.1.0
pandas==2.2.3
plotly==5.24.1
protobuf==5.29.3
PuLP==2.9.0
python-dotenv==1.0.1
sentence_transformers==3.3.1
utilsforecast==0.2.10
dash-mantine-components==0.12.1
google-cloud
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-bigquery[pandas]
google-cloud-secret-manager
google-cloud-storage
# redis
flask==2.2.5
werkzeug==2.2.3
gunicorn>=20.0